requirements.txt
git+https://github.com/bifrost-platform/bifrost-python-lib.git
requests
matplotlib
numpy
python-dotenv